refactor(daemon): split server.rs into server/ with per-domain handler modules
Move blob, tag, ring, and receive handlers into dedicated files under
server/handlers/ to keep the server entry point focused on connection
lifecycle and dispatch.
